Add a way to launch with the menubar hidden & normal window
I prefer to launch the application with a normal window, yet, the
menubar on auto-hide, because most of the time, it's just taking up
space and serves no purpose. This patch introduces a `config.menubar`
setting that makes this possible. Defaults to `true` to preserve the
current behaviour, and is only used when `normal_window` is also used.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>